Fixed logger
